# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  service.CreateOrganizationRequest:
    properties:
      description:
        type: string
      display_name:
        maxLength: 200
        type: string
      domain:
        maxLength: 100
        type: string
      name:
        maxLength: 100
        minLength: 1
        type: string
    required:
    - display_name
    - domain
    - name
    type: object
...
  service.OrganizationListResponse:
    properties:
      organizations:
        items:
          $ref: '#/definitions/service.OrganizationResponse'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  service.OrganizationResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      display_name:
        type: string
      domain:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  service.UpdateOrganizationRequest:
    properties:
      description:
        type: string
      display_name:
        maxLength: 200
        type: string
    required:
    - display_name
    type: object
...
host: localhost:7008
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: This is the backend API for the Developer Portal, providing endpoints
    for managing organizations, teams, projects, components, and deployments.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Developer Portal Backend API
  version: "1.0"
paths:
...
  /organizations:
    get:
      consumes:
      - application/json
      description: Get all organizations with pagination support
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Number of items per page
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved organizations
          schema:
            $ref: '#/definitions/service.OrganizationListResponse'
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List all organizations
      tags:
      - organizations
    post:
      consumes:
      - application/json
      description: Create a new organization with the provided details
      parameters:
      - description: Organization data
        in: body
        name: organization
        required: true
        schema:
          $ref: '#/definitions/service.CreateOrganizationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Successfully created organization
          schema:
            $ref: '#/definitions/service.OrganizationResponse'
        "400":
          description: Invalid request body
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Organization already exists
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a new organization
      tags:
      - organizations
  /organizations/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an organization by ID
      parameters:
      - description: Organization ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Successfully deleted organization
        "400":
          description: Invalid organization ID
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Organization not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete organization
      tags:
      - organizations
    get:
      consumes:
      - application/json
      description: Get a specific organization by its UUID
      parameters:
      - description: Organization ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved organization
          schema:
            $ref: '#/definitions/service.OrganizationResponse'
        "400":
          description: Invalid organization ID
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Organization not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get organization by ID
      tags:
      - organizations
    put:
      consumes:
      - application/json
      description: Update an existing organization by ID
      parameters:
      - description: Organization ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Updated organization data
        in: body
        name: organization
        required: true
        schema:
          $ref: '#/definitions/service.UpdateOrganizationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated organization
          schema:
            $ref: '#/definitions/service.OrganizationResponse'
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Organization not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update organization
      tags:
      - organizations
  /organizations/by-name/{name}:
    get:
      consumes:
      - application/json
      description: Get a specific organization by its name
      parameters:
      - description: Organization name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved organization
          schema:
            $ref: '#/definitions/service.OrganizationResponse'
        "400":
          description: Invalid organization name
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Organization not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get organization by name
      tags:
      - organizations
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"